Fix libvirt/qemu logs location
Fix bind-mount to point /var/log/containers/libvirt/(qemu)
on the host, via the shared NovaLibvirtLogging attribute.

Remove the nested libvirt/qemu path bind-mount since the outer libvirt
directory is sufficient to hold the qemu logs as well. For that,
bind-mount it with the shared mounts propagation mode.

Related: rhbz#2005340
Depends-On: https://review.opendev.org/811096
Change-Id: Iccbae5589aa4a4fc449f23b82755dea1f54f5678
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>