Implement project personas in custom cinder policy file
This commit updates the default cinder policies in
enable-secure-rbac.yaml to implement consistent support for project
personas (project-admin, project-member, and project-reader) with other
OpenStack services. The project-admin is still considered a system
administrator.

This behavior will change in future releases (likely Yoga) when we have
a community wide goal to finish system-scope adoption. At that point, we
can remove these overrides and use the defaults in cinder, which should
adhere to system and project personas and we'll get more functionality
without these overrides.

Change-Id: Ic8199cb3eaa23476157be0f3b69d2ee08f77f17b