Implement project personas in custom cinder policy file
This commit updates the default cinder policies in
enable-secure-rbac.yaml to implement consistent support for project
personas (project-admin, project-member, and project-reader) with other
OpenStack services. The project-admin is still considered a system
administrator.

This behavior will change in future releases (likely Yoga) when we have
a community wide goal to finish system-scope adoption. At that point, we
can remove these overrides and use the defaults in cinder, which should
adhere to system and project personas and we'll get more functionality
without these overrides.

NOTE(stable/wallaby):
Changes are required in order to account for differences between the
wallaby and xena releases. The following policies are removed because
they were added in xena:
  group:group_types:create
  group:group_types:update
  group:group_types:delete
  group:group_types_specs:get
  group:group_types_specs:get_all
  group:group_types_specs:create
  group:group_types_specs:update
  group:group_types_specs:delete
  volume_extension:volume_image_metadata:show
  volume_extension:volume_image_metadata:set
  volume_extension:volume_image_metadata:remove
  volume_extension:volume_type_access:get_all_for_type
  volume_extension:type_create
  volume_extension:type_update
  volume_extension:type_delete
  volume_extension:types_extra_specs:read_sensitive
  volume_extension:quota_classes:get
  volume_extension:quota_classes:update

The following policies are added because they were the deprecated
in xena, but are present in wallaby:
  group:group_types_manage
  group:group_types_specs
  volume_extension:quota_classes
  volume_extension:types_manage
  volume_extension:volume_image_metadata

Co-authored-by: Alan Bishop <abishop@redhat.com>

Change-Id: Ic8199cb3eaa23476157be0f3b69d2ee08f77f17b
(cherry picked from commit 3d2fec12b9f8d95ce3291584d7dc43821774822d)