Implement project personas in custom neutron policy file
This commit updates the default neutron policies in
enable-secure-rbac.yaml to implement consistent support for project
personas (project-admin, project-member, and project-reader) with other
OpenStack services. The project-admin is still considered a system
administrator.

This behavior will change in future releases (likely Yoga) when we have
a community wide goal to finish system-scope adoption. At that point, we
can remove these overrides and use the defaults in neutron, which adhere to
system and project personas and we'll get more functionality without these
overrides.

Change-Id: Ie2273adbdf0dda7947e570933aaca661aa8a67da