Implement project personas in custom barbican policy file
This commit updates the default barbican policies in enable-secure-rbac.yaml to
implement consistent support for project personas (project-admin,
project-member, and project-reader) with other OpenStack services. The
project-admin is still considered a system administrator.

This behavior will change in future releases (likely Yoga) when we have
a community wide goal to finish system-scope adoption. At that point, we
can remove these overrides and use the defaults in barbican, which is working
to adopt full support for system peronas in Yoga, and we'll get more
functionality without these overrides.

Change-Id: Iadd24554727d9a5000c6842954fe1deede833a8e