Fix swift log redirection
Looks like swift log forwarding is skipped atm. Conditions on
ansible blocks are evaluated per task and once a var is
registered, 'is not defined' check (from block) fails and the tasks
are skipped.

This keeps the tasks same in both services so the duplicates
are excluded in DeployStepsTasks.

Also uses shell command like haproxy to check service status.

|         OK | Check if rsyslog exists | overcloud-controller-0
|     TIMING | Check if rsyslog exists | overcloud-controller-0 | 0:15:17.829966 | 0.37s
|       TASK | Forward logging to swift.log file
|    SKIPPED | Forward logging to swift.log file | overcloud-controller-0
|     TIMING | Forward logging to swift.log file | overcloud-controller-0 | 0:15:17.881102 | 0.04s
|       TASK | Restart rsyslogd service after logging conf change
|    SKIPPED | Restart rsyslogd service after logging conf change | overcloud-controller-0
|     TIMING | Restart rsyslogd service after logging conf change | overcloud-controller-0 | 0:15:17.924588 | 0.03s

Change-Id: I2f17b993029fc02ffe1e9d0f6a9c31b9eb97cc0a