Add parameters for manila storage availability zones
Manila supports the configuration of storage availability zones
for each share back end in the cloud. Add
ManilaStorageAvailabilityZone parameter that configures manila's
DEFAULT/storage_availability_zone.

Adds the corresponding Manila<>AvailabilityZone parameter for
each of the manila back ends on this repository. This parameter
is not mandatory, and it defaults to the one configured in the
DEFAULT session of the Manila configuration file.

Depends-On: Ic71545d786b851e7d3f3ce8502e913b81fc70059
Change-Id: I78a64e779d1e55e6ee8de5004457dce7c7906504