Refactor usage of the NovaSchedulerEnabledFilters parameter
The NovaSchedulerEnabledFilters parameter accepts a list of scheduler
filters enabled. This parameter is currently used with a list defined
in single line but this makes it difficult to understand diff when we
add/remove items to/from the list.

This change expands the list value so that any addtion/removal results
in a clean one line diff.

Change-Id: Ib5e9e4aeb100ea347a354878c56e45dfe97bf999