Add templates for custom network ci job
We are creating a new featureset064 that will deploy overcloud with
custom network + custom overcloud name + IPA integration.

With this patch, Adding tht templates that we will use in fs064 for
overcloud deployment.

* ci/custom_ci_roles_data.yaml file have some changes compared to
master branch:-
CinderHPELeftHandISCSI was added as this was present in wallaby
ManilaBackendFlashBlade was removed as this was not present in wallaby

Depends-On: https://review.opendev.org/c/openstack/tripleo-ansible/+/821170
Change-Id: I5119a2c0103868fd481b57b890947b2881b86845
(cherry picked from commit 2e4f35b07d39ae0d959a16a49bf3bfa3fbbe8fa4)