Add InternalApiNetwork entry in ServiceNetMap
This is not explicitly mapping to a network to a service,
but we already have other entries like ResolveNetwork and
PublicNetwork in the ServiceNetMap.

Adding InternalApiNetwork to the ServiceNetMap will allow us
to remove a conditional on hard-coded network name in
tripleo_hieradata.

It also makes it possible to change the network for interal_api_vip
via the ServiceNetMap or the service_net_map_replace in network
data.

Conflicts:
  overcloud-resource-registry-puppet.j2.yaml

Related-Bug: #1946239
Change-Id: I1dcd9da59487bfa0842c5f131a0e9a9d9d0c676d
(cherry picked from commit 3cf046ccee8e2c90a9b03fd6bb32865016fa8a24)