Flatten yaql list expressions
In python yaql 2, the expressions are returning a nested list but we
need a list of dictionaries in order to use map_merge.  This change adds
flatten so it returns [] instead of [[]] when nothing is available.

Change-Id: Ic5144c58ceb9bd146e2c470725ec7f4b65328c4d
Closes-Bug: #1947193
(cherry picked from commit 6b50e162235f083163e22e1f7dd9157e6c859e94)