Flatten yaql list expressions
In python yaql 2, the expressions are returning a nested list but we
need a list of dictionaries in order to use map_merge.  This change adds
flatten so it returns [] instead of [[]] when nothing is available.

Change-Id: Iba0262fbc2af0903f02c8c28c7ba2b1d935abe7f
(cherry picked from commit 20311a0a50af684f2eab1a98adb58fe94f5ce933)