Fix set_fact on the CephMgrAnsibleVars
The purpose of this patch is react on the issue generated by [1],
where an empty dict sect_fact is not allowed for ansible >= 2.11.
For this reason, we're going to always set those variables and
defer to the ansible layer the evaluation of the DashboardEnabled
condition.

[1] https://github.com/ansible/ansible/pull/73684

Closes-Bug: #1949033
Change-Id: I4ab2d20df22bdc9b140019f90d1370ef17f09294
(cherry picked from commit b3863568ad75a3845389677b08a0d756e3c6e0b5)