[TRAIN-ONLY] Add a new upgrade task for some services
Those services are mostly managed as bundles by pacemaker. Since we
can't backport Ic3e427156fc82c524c763d1896937fcc3c49fabb down to
stable/train[1], we have to tweak an upgrade task that will ensure we have
the correct setype.

This is the complement of another train-only patch merged lately:
https://review.opendev.org/c/openstack/tripleo-heat-templates/+/808964

Those "new" locations weren't detected before, because it seems to fail
randomly during the FFU. First it was OVN, then, maybe, HAProxy. Or
something else.

Note:
cinder is NOT affected by this patch, since it's a really complicated
service, especially when NFS comes in the game. We'd rather get some
documentation for this specific service FFU, instead of risking a
complete crash.

[1] reason: since the pacemaker bundle would be updated, this would lead
to a complete service outage on any day-2 operation, being a simple
scale-out for instance. We obviously can't do that. There might be a way
to avoid such an outage, but the modification involved would make the
overall bundle definition more complicated, more error-prone, and the
actual benefit would be ridiculous compared to the risks.

Change-Id: If796d3c1d37c72655bd6b99e851a3b3e7617b9a4