Add rgw max attributes constraints
When the swift /info endpoint is queried, Tempest fails checking the
following missing keys:

1. max_meta_count
2. max_meta_name_length
3. max_meta_value_length

Those keys are not present because the related rgw settings are missing,
and as per [1], nothing is generated if the rgw_max_* are not set.
This patch is a workaround, on the TripleO side, to provide the missing
keys to the cluster.

[1] https://github.com/ceph/ceph/blob/master/src/rgw/rgw_rest_swift.cc#L1883-L1900

Closes-Bug: #1950486
Change-Id: I0b199e47cce9dc72490110105f6a4ff94f90934d