Check that redis is removed prior to FFU
We want to disable Redis by default in new deployments, without
impacting upgrade of existing deployments.

Add a pre-check during upgrade and bail out if Redis is not
enabled explicitely during upgrade and if the resource already
exists in pacemaker. The pre-check is performed prior to
os upgrade so that pacemaker is still installed and the cluster
can still be inspected.

Tested as follows:
  . FFU with redis-bundle not removed fails as expected
  . FFU with redis-bundle removed finishes
  . FFU when redis service is enabled post-FFU always finishes
  . re-running os-upgrade is idempotent

Change-Id: Icd7503c9b9b5edc7ecfe20e6217fb8efeebf3140
Closes-Bug: #1872645