Ensure we bind-mount /dev instead of /dev/
With the move to crun instead of runc for the container engine, we seem
to hit a known issue that was corrected back in 2015[1] for runc. There
was then a regression, fixed with [2] a bit later.

There's a good chance crun has a partial fix only, matching only /dev
and not /dev/, leading to the change of /dev/ptmx from an actual node to
a symlink pointing to /dev/pts/ptmx.

Another fix might be ensuring we don't have any trailing "/" in the
volume paths passed to the tripleo-ansible/tripleo_container_manage
module/role.

[1] https://github.com/opencontainers/runc/pull/96
[2] https://github.com/opencontainers/runc/pull/742/files

Closes-Bug: #1950176

Change-Id: I094120f7f2f6bfcfc0cc5843aa1b23629cd90a23