Introduce a new linter for yaml-validate, and correct issues
This new linter ensures we don't have any trailing "/" in the container
volume definitions.

Those trailing "/" may create issues with the containers, for instance
for specific mounts such as "/dev"[1].

This patch also takes the opportunity to fix those trailing "/" for the
affected files, in order to start on a clean basis.

[1] https://launchpad.net/bugs/1950176

Change-Id: If951f9643d67574c1225301aab7c9e4b0d316b7f
Related-Bug: #1950176