Add CephAdmVerbose variable
The final step of the cephadm playbook calls post.yaml, which is able
to print the status of the cluster and export it as a ceph compatible
spec retrieved from the Ceph cluster. However, in a tls-everywhere
context, we should be able to stop logging the spec as it contains
sensitive info (the priv key).
This change just adds the CephAdmVerbose tht varibale to control what
is printed. The variable is set to true in both Ceph related scenarios
tested in the tripleo-ci.

Depends-On: I7bef5922d75114e4bd1924b52a382e61d7027d2d
Change-Id: I1db48dcc004223b90c3505bd447530593be487bf