[train,ussuri] Set cinder's __DEFAULT__ vol type description
Update the description of cinder's internal __DEFAULT__ volume type
so that users don't confuse it with the actual default volume type
established by the CinderDefaultVolumeType parameter.

This patch is a variant of Idf27c14b31dc077ef9a0e567bd502ed6842bd52b,
which in later releases deletes cinder's __DEFAULT__ volume type if
there are no volumes associated with it. However, the __DEFAULT__
volume type cannot be removed prior to the victoria release because
it's required for a DB migration.

Lastly, CinderDefaultVolumeType is now constrained to prevent it being
set to an empty string. That should never happen, so this is just a
safety net.

Change-Id: Ifc35eb09ab3af4d8f3ea232dd20558047c6834b9
Related-Bug: #1782217