Add chrony waitsync back in
This is a revert of 373a27163e0365919f707604df5b61863029d2e4 because
makestep does not fail if the ntp servers are unavailable. This change
adds waitsync back in so it fails if the servers are unavilable but
leaves the makestep in place which should force the sync.  If the
servers are unavailable, the waitsync should catch it. If they are
available waitsync should be a relative noop because the makestep would
sync the system.  The waitsync has been increased to 5 minutes (from 200
seconds). waitsync will wait 30 times with an interval of 10 seconds
between checks.

Closes-Bug: #1955414
Change-Id: I4ac9210f6533e6826b2814f72b7271b43fdca267