Enable policy rule management in nova-compute
There are some features in nova-compute which validate policy rules.
For example when connecting an instance to an external network,
nova-compute checks ``network:attach_external_network`` to determine
whether the operation is permitted.

This change migrates definition of nova::policy hieradata from nova-api
to nova-base so that the hieradata can be used by both nova-api and
nova-compute.

Closes-Bug: #1955786
Depends-on: https://review.opendev.org/822991
Change-Id: I90be98d67ede06933bbcf3ad5448365ea220f1a9