Pass OS_CLOUD env when migrating to heat ephemeral.
In order for the undercloud-upgrade-ephemeral-heat.py to work properly
we need to specify the OS_CLOUD for the existing undercloud stack,
otherwise the invoked OpenStack commands won't be able to authenticate
properly.

This change makes use of the --cloud script's option to pass it into
each subprocess' invokation of an openstack command.

Change-Id: I0b05a05907c42abfabe9cb963da65ed1efb2d283
Resolves: rhbz#2040260