[train-only][ffwd][allinone] Fix idempotency for all in one upgrade
Before running leapp on the nodes we need to pause ceph as
the whole ceph cluster will go down. This has to be done just
once but if we fail somewhere in system upgrade - wrong leapp
repos for example - we end up in situation where we don't know
if ceph is paused or not and we can't check as caph can be
partially or fully down. Best way is to make sure we create state
file right after succesful pause on the node where we executed
the operation.

Change-Id: I70ba5016a63a08a1bf2650a4e24f889b24c96f44