Nova: Use internal endpoint to access Keystone
The Keystone admin endpoint is deployed on the provisioning network by
default, and usually this network has no redundancy. We don't expect
outage in provisioning network affects functionality of overcloud so
services should not be rely on connectivity in provisioning network.

This change ensures that internal endpoint, which is deployed on
internal api network, is used instead.

Closes-Bug: #1959353
Change-Id: Id13a221726af88dd229cf8e12462eabb4027b377