Use Internal API by default for Keystone admin endpoint
We have been using ctlplane network by default for Keystone admin
endpoint. The intention was to let undercloud talk with the endpoint to
create resources like endpoints, but that requirement is not true,
since we use tripleo-ansible, which relies on public endpoint, to
create keystone resources.

This change updates the default to internal api network. The ctlplane
network, which is the provisioning network, usually has no redundancy,
and we should not make any of overcloud functionalities rely on this
network if there is no actual requirement to do so.

Closes-Bug: #1959459
Change-Id: I931e4c3c8a081adb144d915410e0d38760466561