Disable heat by default in favor of ephemeral-heat.
Now that the default way to deploy heat is using the ephemeral-heat
service, we are not required to pass a specific environment file to
ensure the service gets disabled for the Undercloud.

This patch moves the content from disable-heat.yaml environment file
into the services default's for the undercloud with the intention of
stopping passing it by default when enable_heat is not passed.

Change-Id: I1419c1bb901b3fb7ca038c941c230691aff2ebbd
(cherry picked from commit ec33ab40488efacaf98e31852b280947ee73bff8)