Add parameter IronicIPXEUefiSnpOnly
Add new parameter IronicIPXEUefiSnpOnly (default: true).
When `true` ipxe-snponly.efi is used for UEFI, when `false` ipxe.efi
is used.

'snponly.efi' is the default in puppet-ironic master branch, however
in Wallaby and earlier releases 'ipxe.efi' is the default. Since
'ipxe.efi' is not compatible with a lot of hardware TripleO should
change this default.

Related: RHBZ#2049179
Closes-Bug: 1959726
Change-Id: I8ee338c3f3e20f1826d98efb38d3b21fefda5031