[train-only][ffwd][hybrid-mode] Hardcode neutron rpc_response_timeout to 60 on hybrid mode
During hybrid mode we often hit case when rpc_response_timeout 10+min
which would be fine value of big environments but neutron does not
have heartbeat. This means that by the time control plane gets up
we might end up with 10min of outage before the neutron agents connect.

There is expectation that during hybrid mode only live migrations are
done so low value should be safe enough. This will get reset to
whatever value was set in templates on the node upgrade.

Closes-bug: rhbz#2049690

Change-Id: Ibd21764992f401d06b24718ea28cb2735de53858