Fix OS_CLOUD for multistack and nova az tasks
The entry in clouds.yaml will correspond to the stack that deploys
keystone. Therefore in a multistack deployment RootStackName will not be
give the correct entry. Add a parameter to override this.

Also clouds.yaml is owned by root and mode 0600 so we much use
become: true.

Update nova az tasks to use clouds.yaml.

Change-Id: I7e97108d8165c16946b3b11f3084bfd51e3128dc