Skip podman purge on Undercloud
With the undercloud using ephemeral Heat, there will be images for
ephemeral Heat in the local container storage that do not have any
associated containers. Podman purge would delete these images, and they
then wouldn't be available for overcloud commands.

Instead, skip running the podman purge tasks on the Undercloud. This
means that the local container image storage will grow over time. Given
the reduced number of services on the undercloud, the storage usage is
not as bad as it was previously. However, operators will need to be
aware.

While "podman image prune" does have a --filter option to preserve
images, it does not work in the version of podman in
container-tools:3.0. When the needed patch[1] is available, the podman
purge tasks can be re-enabled, which would address the issue of growing
local storage.

[1] https://github.com/containers/podman/commit/a5ad36c65ea07d839fd9bf55a820c8cb9884eed1
[2] https://github.com/containers/podman/pull/10390

Change-Id: Ibd2200aa912960b267a9975571e05665a29e40f0
Signed-off-by: James Slagle <jslagle@redhat.com>