Properly set ceph-ansible extra_vars and tags for cephadm-adopt
When cephadm-adopt tag is passed, both external_upgrades and external
_deploy steps are run. However, ceph-ansible extra vars and skip tags
should be set once (during the external upgrade execution). This patch
ensures that during the cephadm deploy steps these two tasks are not
executed when the cephadm_adopt tag is passed (as this tag is not part
of their definition anymore).

Change-Id: Ie34617745f4312496a26173c4ba043cb7e473663