Redis: Hard-code paths of configuration files
In current CentOS 9 Stream, configuration files of redis is placed in
/etc/redis. Following that, the puppet-redis module has been updated to
use the new path by [1].

This change ensures that the old path is consistently used in both
CentOS 8 and CentOS 9 as we require a hard-coded path to start up
redis in standalone mode(without pacemaker). The path can be updated
later, for example when we drop support for CentOS 8.

[1] https://github.com/voxpupuli/puppet-redis/pull/434

Closes-Bug: #1960620
Change-Id: Ibf7c2f8d01ddf73c92d183e51c378a0770002e72