Expose tripleo_cephadm_default_container boolean
Even if Ceph is deployed before Heat runs, the Ceph
Dashboard downloads Ceph containers. If the user opts
to get Ceph contianers based on the default in cephadm
and not from container_image_prepare_defaults.yaml,
then the same choice should be applied during the
overcloud deployment when the dashboard is deployed.

The deployed_ceph.yaml environment file generated
by 'openstack overcloud ceph deploy' will set this
parameter this patch introduces so the user will not
need to set it more than once.

Depends-On: Ic4f626ebbe1a5332f9f285e5ea6aaa9f24a400b8
Change-Id: I0f64c8738d65c08561a22b09dc9f5ed48af0935c