Update of OVN controllers as an external task.
OVN requires ovn_controller to be updated before ovn_northd.

This implement a new step in the update process to be run before
"overcloud update run ..." commands and after "overcloud update
prepare".

Running:

  openstack overcloud external-update run --stack qe-Cloud-0 --tags ovn

triggers an update of the ovn_controller containers on every
server in parallel.

We don't add the whole configuration of the ovn_controller here,
meaning we *don't* run puppet for the new ovn_controller. Maybe this
should be added as an option, but the time taken by this step will
increase. Current testing show that it's not mandatary to have a
successful update.

Note that new configuration will be taken into account later during
the "update run" commands.

By doing this we avoid any issues where a discrepency in version
between ovn-northd and ovn-controller would cause a cut in the
data-plane.

Change-Id: Icc58fa5f3ccfea9b12154629be977504fe97b45f