THT cleanup of scenario001/004/010 for Deployed Ceph
Remove parameters which are not used when "deployed ceph"
is used. The CephClusterFSID, "DeployedCeph: true", and
data_devices (from the soon to be deprecated CephOsdSpec)
will be set in the deployed_ceph.yaml environment file
which is generated and passed to the overcloud deployment
by I982dedb53582fbd76391165c3ca72954c129b84a.

Change-Id: Id6d0699123b9e9b75011ee94949ec4a69012f9af