Disable auto discovery
Tripleo-ci jobs are broken after latest release of setuptools 61.0
because of breaking changes which are not backwork compatible,
details in related bug and [1].

Users that don't set ``packages``, ``py_modules``, or `configuration`
are still likely to observe the auto-discovery behavior, which may
halt the build if the project contains multiple directories and/or
multiple Python files directly under the project root.

To disable auto discovery, one can do below in setup.py

~~~
setuptools.setup(..,packages=[],..)
~~~

or

~~~
setuptools.setup(..,py_modules=[],..)
~~~

The Depends-On is irrelevant to the bug but is needed for job freeze
issue seen on this review.

[1] https://github.com/pypa/setuptools/issues/3197

Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/835776
Related-Bug: #1966382
Change-Id: If1509b0369aeaee72355da4dd3a5cdd7846114bb
(cherry picked from commit 0176edf256e69c266845194e57882c9e9cb5ef94)