Designate: configure collocation IP on bind and neutron dns integration
This patch moves the collcation IP that was on unbound to the bind
worker instance allowing the use of the framework provided to configure
the neutron services to refer to the designate DNS services.

Note: the ansible tasks were move to the bind config ansible role
instead of the template.

Depends-On: https://review.opendev.org/c/openstack/tripleo-ansible/+/839973
Depends-On: https://review.opendev.org/c/openstack/puppet-tripleo/+/835742
Change-Id: Ia6a8128b0c87979fd118a7ae8cd33e0430a4aad1