Set rx permissions for all in /var/lib/config-data
When using common-container-setup-tasks.yaml after
deployment/unbound/unbound-container-ansible.yaml the permission for
config-data are set to 750 since they are not set explicitely in
common-container-setup-tasks.yaml. This caused problems due to other
users not being able to read anythin below that folder. This changes
sets explicitely the permissions to 755 to avoid these problems.

Change-Id: I54fad8eb65bdcfd5d71a9999f2fb9ff97c143a62