Remove ineffective ovn_remote_probe_interval
This parameter was introduced to puppet-neutron but it was completely
wrong and the actual parameter has never been implemented in neutron.

Also, the parameter definition in tht has never taken effect because
puppet-tripleo overrides the definition.

The interval should be configured in ovn controller, and this is
already implemented using puppet-ovn.

Change-Id: I36499dccef6ff74a288942bd0a8f0257187387e7