Add AdminPassword and KeystoneRegion stack outputs
AdminPassword and KeystoneRegion need to be added as a stack outputs so
that they are saved in the working directory for stack outputs after
using ephemeral Heat to create the stack.

The code to create the rc params to create the overcloudrc file(s)
(tripleoclient.utils.get_rc_params) can then create the overcloudrc
using only the working directory. No reference to the stack object or a
running instance of Heat to query the stack would be required.

Change-Id: Idaef781163c6c8f5928d93d9bbc1aa7b0dee6fd6
Signed-off-by: James Slagle <jslagle@redhat.com>
(cherry picked from commit 4d91ad67fc17a0bc351a48dd1187109c3ae96ca7)